+++ EssentialApp/EssentialAppTests/FeedAcceptanceTests.swift
@@ -76,31 +76,6 @@ class FeedAcceptanceTests: XCTestCase {
         sut.sceneWillResignActive(UIApplication.shared.connectedScenes.first!)
     }
     
-    private class HTTPClientStub: HTTPClient {
-        private class Task: HTTPClientTask {
-            func cancel() {}
-        }
-        
-        private let stub: (URL) -> HTTPClient.Result
-        
-        init(stub: @escaping (URL) -> HTTPClient.Result) {
-            self.stub = stub
-        }
-        
-        func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> HTTPClientTask {
-            completion(stub(url))
-            return Task()
-        }
-        
-        static var offline: HTTPClientStub {
-            HTTPClientStub(stub: { _ in .failure(NSError(domain: "offline", code: 0)) })
-        }
-        
-        static func online(_ stub: @escaping (URL) -> (Data, HTTPURLResponse)) -> HTTPClientStub {
-            HTTPClientStub { url in .success(stub(url)) }
-        }
-    }
-    
     private class InMemoryFeedStore: FeedStore, FeedImageDataStore {
         private(set) var feedCache: CachedFeed?
         private var feedImageDataCache: [URL: Data] = [:]
